final refinements
