+++ activiti-engine/src/test/java/org/activiti/engine/test/api/repository/RepositoryServiceTest.java
@@ -324,7 +324,7 @@ public class RepositoryServiceTest extends PluggableActivitiTestCase {
   public void testGetBpmnModel() {
     ProcessDefinition processDefinition = repositoryService.createProcessDefinitionQuery().singleResult();
 
-    // Some set0 assertions
+    // Some basic assertions
     BpmnModel bpmnModel = repositoryService.getBpmnModel(processDefinition.getId());
     assertNotNull(bpmnModel);
     assertEquals(1, bpmnModel.getProcesses().size());
