+++ activiti-spring-conformance-tests/activiti-spring-conformance-set0/src/test/resources/processes/UserTask with no User or Group Assignment.bpmn20.xml
