+++ activiti-spring-conformance-tests/activiti-spring-conformance-set1/src/test/resources/processes/ServiceTask with Implementation.bpmn20.xml
