+++ activiti-spring-conformance-tests/activiti-spring-conformance-set2/src/test/resources/processes/UserTask with CandidateUser.bpmn20.xml
