+++ activiti-spring-conformance-tests/activiti-spring-conformance-util/src/main/java/org/activiti/spring/conformance/util/security/SecurityUtil.java
